Move demand CSV code into submodule of `input`
Closes #285.
