+++ src/lib.rs
@@ -3,7 +3,6 @@
 pub mod agent;
 pub mod commands;
 pub mod commodity;
-pub mod demand;
 pub mod input;
 pub mod log;
 pub mod model;
